+++ eslint.config.mjs
@@ -1,15 +1,15 @@
-import eslintPluginJs from "@eslint/js";
-import eslintPluginStylistic from "@stylistic/eslint-plugin";
+import {defineConfig, globalIgnores} from "eslint/config";
+import css from "@eslint/css";
 import globals from "globals";
-import {flatConfigs as importConfigs} from "eslint-plugin-import-x";
+import {flatConfigs as importX} from "eslint-plugin-import-x";
+import js from "@eslint/js";
+import json from "@eslint/json";
+import markdown from "@eslint/markdown";
+import stylistic from "@stylistic/eslint-plugin";
 
-const config = [
-  eslintPluginJs.configs.recommended,
-  eslintPluginStylistic.configs.all,
-  importConfigs.recommended,
-  {
-    "ignores": ["**/*.min.js"]
-  },
+export default defineConfig([
+  globalIgnores(["**/*.min.js"]),
+  {"files": ["**/*.css"], "languageOptions": {"tolerant": true}, "plugins": {css}, "language": "css/css", "extends": ["css/recommended"], "rules": {"css/use-baseline": ["error", {"available": "newly"}]}},
   {
     "files": ["**/*.js"],
     "languageOptions": {
@@ -19,9 +19,10 @@ const config = [
         ...globals.node,
         "$item": "writable",
         "$node": "writable"
-      },
-      "sourceType": "commonjs"
+      }
     },
+    "plugins": {js, stylistic},
+    "extends": [importX.recommended, "js/recommended", "stylistic/all"],
     "rules": {
       "@stylistic/array-element-newline": ["error", "consistent"],
       "@stylistic/brace-style": ["error", "1tbs", {"allowSingleLine": true}],
@@ -58,12 +59,17 @@ const config = [
       },
       "sourceType": "module"
     },
+    "plugins": {js, stylistic},
+    "extends": [importX.recommended, "js/all", "stylistic/all"],
     "rules": {
       "@stylistic/array-element-newline": ["error", "consistent"],
       "@stylistic/indent": ["error", 2],
-      "no-magic-numbers": "off"
+      "@stylistic/object-property-newline": ["error", {"allowAllPropertiesOnSameLine": true}],
+      "import-x/no-unresolved": ["error", {"ignore": ["eslint/config"]}],
+      "no-magic-numbers": "off",
+      "sort-keys": "off"
     }
-  }
-];
-
-export default config;
+  },
+  {"files": ["**/*.json"], "ignores": ["package-lock.json"], "plugins": {json}, "extends": ["json/recommended"], "language": "json/json"},
+  {"files": ["**/*.md"], "plugins": {markdown}, "extends": ["markdown/recommended"], "language": "markdown/gfm"}
+]);